refactor: move api into core app
