+++ core/urls.py
@@ -1,4 +1,4 @@
-from django.urls import path
+from django.urls import include, path
 
 from .views import (
     CommunityCreateView,
@@ -20,4 +20,5 @@ urlpatterns = [
     path("communities/", CommunityListView.as_view(), name="community-list"),
     path("community/create/", CommunityCreateView.as_view(), name="community-create"),
     path("community/<slug:slug>/", CommunityDetailView.as_view(), name="community-detail"),
+    path("api/", include("core.api_urls")),
 ]
