+++ reddit/urls.py
@@ -28,7 +28,6 @@ urlpatterns = [
     path("", include("articles.urls")),
     path("users/", include("users.urls")),
     path("core/", include("core.urls")),
-    path("api/", include("api.urls")),
 ]
 
 if settings.DEBUG:
